+++ .eslintignore
@@ -2,3 +2,4 @@ lib
 decls
 node_modules
 spec/fixtures
+vendor/commands.js
