+++ package.json
@@ -15,7 +15,11 @@
     "sb-stream-promise": "^1.0.0"
   },
   "scripts": {
-    "test": "apm test"
+    "lint": "(flow check) && (eslint .)",
+    "test": "apm test",
+    "clean": "rm -rf lib",
+    "compile": "npm run clean && babel src --out-dir lib",
+    "watch": "npm run clean && babel src --out-dir lib --watch"
   },
   "repository": {
     "type": "git",
